重构t_operation表

+++ cat-core/src/main/java/com/dianping/cat/core/mybatis/operation/dao/OperationMapper.java
@@ -7,9 +7,9 @@ import org.apache.ibatis.annotations.Param;
 import com.dianping.cat.core.mybatis.operation.dao.data.OperationDO;
 
 public interface OperationMapper {
-	int deleteByPrimaryKey(@Param("id") Integer id);
+	int deleteByPrimaryKey(@Param("id") Long id);
 
-	OperationDO findByPrimaryKey(@Param("id") Integer id);
+	OperationDO findByPrimaryKey(@Param("id") Long id);
 
 	int insert(OperationDO record);
